Ecomm_Backend/requirements.txt
alembic==1.8.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.0
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
PyJWT==2.5.0
pyparsing==3.0.9
six==1.16.0
SQLAlchemy==1.4.42
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.10.0
PyMySQL==1.0.2
